+++ src/components/gallery/Gallery.js
@@ -1,9 +1,9 @@
 import React, { useEffect, useState } from "react";
-import axios from "axios";
 import GalleryCard from "./GalleryCard";
-import {Error} from "./layout/ErrorElements"
-import {PageTitle} from "./layout/PageElements"
-import { GalleryPageContainer, CardContainer } from "./layout/GalleryElements";
+import {Error} from "../layout/ErrorElements"
+import {PageTitle} from "../layout/PageElements"
+import { GalleryPageContainer, CardContainer } from "../layout/GalleryElements";
+import {dataHandler} from "../data_handler";
 
 const Gallery = (props) => {
   // const url = "https://api.nasa.gov/planetary/apod?api_key=DEMO_KEY&count=25";
@@ -13,18 +13,7 @@ const Gallery = (props) => {
   const [error, setError] = useState("");
 
   useEffect(() => {
-    axios
-      .get(url)
-      .then((res) => {
-        setGallery(res.data);
-      })
-      .catch((error) => {
-        setError(error.message);
-        console.error(
-          `The request was made and the server responded
-        with a status code that falls out of the range of 2xx ` + error.message
-        );
-      });
+    dataHandler._api_get(url, setGallery, setError);
   }, [url]);
 
   return (
