+++ src/components/gallery/GalleryCard.js
@@ -6,12 +6,12 @@ import {
   Date,
   CloseButton,
   HdImg,
-} from "./layout/GalleryElements";
+} from "../layout/GalleryElements";
 import {
   faTimes
 } from "@fortawesome/free-solid-svg-icons";
 import Modal from "react-modal";
-import LoadingAnim from "../static/img/loading.gif";
+import LoadingAnim from "../../static/img/loading.gif";
 
 // Modal.setAppElement("#root");
